Move things to rsdp_search.rs
